Add files via upload
The idea for us will be to determine how temperature changes over time. So that automatically leads to a derivative of temperature with respect to time, which will basically boil down to a differential equation. The variation of temperature of this model is given by the following linear affine differential equation.

The constant a.
The decay rate of a temperature.
Heater capacity
Heater switch(state 1 or 0)